(refactor) moved markets category to component
