+++ app/src/components/common/icons/IconFilter.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export const IconFilter = () => (
+  <svg fill="none" height="20" viewBox="0 0 20 20" width="20" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M17.5 9.16667H5.69167L8.675 6.175L7.5 5L2.5 10L7.5 15L8.675 13.825L5.69167 10.8333H17.5V9.16667Z"
+      fill="#37474F"
+    />
+  </svg>
+)
