+++ app/src/components/market/market_home.tsx
@@ -1,13 +1,10 @@
-import { useInterval } from '@react-corekit/use-interval'
-import React, { useCallback, useEffect, useState } from 'react'
-import Draggable from 'react-draggable'
+import React, { useEffect, useState } from 'react'
 import styled from 'styled-components'
 
 import { ConnectedWeb3Context } from '../../hooks/connectedWeb3'
 import { RemoteData } from '../../util/remote_data'
-import { Button, ButtonCircle, ButtonSelectable, ListCard, ListItem, Loading, SectionTitle } from '../common'
-import { IconChevronLeft } from '../common/icons/IconChevronLeft'
-import { IconChevronRight } from '../common/icons/IconChevronRight'
+import { Button, ButtonSelectable, ListCard, ListItem, Loading, SectionTitle } from '../common'
+import { MarketsCategories } from '../common/markets_categories'
 
 const CATEGORIES = ['All', 'Politics', 'Cryptocurrencies', 'Sports', 'Esports', 'NBA']
 
@@ -17,35 +14,6 @@ const SectionTitleMarket = styled(SectionTitle)`
 
 const TopContents = styled.div`
   padding: 25px;
-`
-
-const CategoriesWrapper = styled.div`
-  align-items: center;
-  border-bottom: 1px solid ${props => props.theme.borders.borderColor};
-  display: flex;
-  margin-bottom: 20px;
-  padding-bottom: 20px;
-  width: 100%;
-`
-
-const CategoriesButtons = styled.div`
-  flex-grow: 1;
-  overflow: hidden;
-`
-
-const CategoriesButtonsInner = styled.div`
-  align-items: center;
-  display: flex;
-  width: fit-content;
-`
-
-const CategoriesControls = styled.div`
-  column-gap: 10px;
-  display: grid;
-  flex-grow: 0;
-  flex-shrink: 0;
-  grid-template-columns: 1fr 1fr;
-  margin-left: 15px;
 `
 
 const SelectableButton = styled(ButtonSelectable)`
@@ -66,10 +34,15 @@ const FiltersButtons = styled.div`
   display: flex;
 `
 
+const ListWrapper = styled.div`
+  display: flex;
+  min-height: 200px;
+  flex-direction: column;
+`
+
 const NoMarketsAvailable = styled.p`
-  align-self: center;
   font-size: 18px;
-  margin: 0;
+  margin: auto 0;
   text-align: center;
 `
 
@@ -83,21 +56,11 @@ interface Props {
   onShowMore: () => void
 }
 
-enum SliderDirection {
-  left,
-  right,
-  none,
-}
-
 export const MarketHome: React.FC<Props> = (props: Props) => {
   const { context, count, markets, moreMarkets, onFilterChange, onShowMore } = props
   const [state, setState] = useState('OPEN')
   const [category, setCategory] = useState('All')
   const [sortBy, setSortBy] = useState<Maybe<string>>(null)
-  const [sliderXDisplacement, setSliderXDisplacement] = useState<number>(0)
-  const [sliderMoving, setSliderMoving] = useState<SliderDirection>(SliderDirection.none)
-  const [sliderButtonDisabled, setSliderButtonDisabled] = useState<SliderDirection>(SliderDirection.left)
-  const [sliderRange, setSliderRange] = useState<number>(0)
 
   useEffect(() => {
     onFilterChange({ category, sortBy, state })
@@ -109,104 +72,19 @@ export const MarketHome: React.FC<Props> = (props: Props) => {
     </Button>
   ) : null
 
-  const categoriesButtonsRef: any = React.createRef()
-  const categoriesButtonsInnerRef: any = React.createRef()
-  const X_DISPLACEMENT = 5
-  const DISPLACEMENT_TIMER = 5
-
-  const resetSliderButtons = useCallback(() => {
-    if (sliderButtonDisabled !== SliderDirection.none) {
-      setSliderButtonDisabled(SliderDirection.none)
-    }
-  }, [sliderButtonDisabled])
-
-  const sliderShouldMove = useCallback(
-    (sliderDirection: SliderDirection) => {
-      return sliderMoving === sliderDirection ? DISPLACEMENT_TIMER : null
-    },
-    [sliderMoving],
-  )
-
-  useEffect(() => {
-    if (categoriesButtonsInnerRef.current.clientWidth < categoriesButtonsRef.current.clientWidth) {
-      setSliderRange(0)
-    } else {
-      setSliderRange(categoriesButtonsInnerRef.current.clientWidth - categoriesButtonsRef.current.clientWidth)
-    }
-  }, [categoriesButtonsInnerRef, categoriesButtonsRef])
-
-  useInterval(() => {
-    const moveResult = sliderXDisplacement - X_DISPLACEMENT
-
-    if (Math.abs(moveResult) > sliderRange) {
-      setSliderButtonDisabled(SliderDirection.right)
-      return
-    }
-
-    resetSliderButtons()
-    setSliderXDisplacement(moveResult)
-  }, sliderShouldMove(SliderDirection.right))
-
-  useInterval(() => {
-    const moveResult = sliderXDisplacement + X_DISPLACEMENT
-
-    if (moveResult > 0) {
-      setSliderButtonDisabled(SliderDirection.left)
-      return
-    }
-
-    resetSliderButtons()
-    setSliderXDisplacement(moveResult)
-  }, sliderShouldMove(SliderDirection.left))
-
-  const cancelSliding = () => {
-    setSliderMoving(SliderDirection.none)
-  }
-
   return (
     <>
       <SectionTitleMarket title={'Markets'} />
       <ListCard>
         {context.account && (
           <TopContents>
-            <CategoriesWrapper>
-              <CategoriesButtons ref={categoriesButtonsRef}>
-                <Draggable
-                  axis="x"
-                  defaultPosition={{ x: 0, y: 0 }}
-                  disabled
-                  position={{ x: sliderXDisplacement, y: 0 }}
-                >
-                  <CategoriesButtonsInner ref={categoriesButtonsInnerRef}>
-                    {CATEGORIES.map((item, index) => (
-                      <SelectableButton active={item === category} key={index} onClick={() => setCategory(item)}>
-                        {item}
-                      </SelectableButton>
-                    ))}
-                  </CategoriesButtonsInner>
-                </Draggable>
-              </CategoriesButtons>
-              {sliderRange !== 0 && (
-                <CategoriesControls onMouseLeave={cancelSliding}>
-                  <ButtonCircle
-                    disabled={sliderButtonDisabled === SliderDirection.left}
-                    onMouseDown={() => setSliderMoving(SliderDirection.left)}
-                    onMouseLeave={cancelSliding}
-                    onMouseUp={cancelSliding}
-                  >
-                    <IconChevronLeft />
-                  </ButtonCircle>
-                  <ButtonCircle
-                    disabled={sliderButtonDisabled === SliderDirection.right}
-                    onMouseDown={() => setSliderMoving(SliderDirection.right)}
-                    onMouseLeave={cancelSliding}
-                    onMouseUp={cancelSliding}
-                  >
-                    <IconChevronRight />
-                  </ButtonCircle>
-                </CategoriesControls>
-              )}
-            </CategoriesWrapper>
+            <MarketsCategories>
+              {CATEGORIES.map((item, index) => (
+                <SelectableButton active={item === category} key={index} onClick={() => setCategory(item)}>
+                  {item}
+                </SelectableButton>
+              ))}
+            </MarketsCategories>
             <FiltersWrapper>
               <FiltersButtons>
                 <SelectableButton active={state === 'OPEN'} onClick={() => setState('OPEN')}>
@@ -229,14 +107,16 @@ export const MarketHome: React.FC<Props> = (props: Props) => {
             </FiltersWrapper>
           </TopContents>
         )}
-        {RemoteData.hasData(markets) &&
-          markets.data.length > 0 &&
-          markets.data.slice(0, count).map(item => {
-            return <ListItem key={item.id} market={item}></ListItem>
-          })}
-        {RemoteData.is.success(markets) && markets.data.length === 0 && (
-          <NoMarketsAvailable title={'No markets available'} />
-        )}
+        <ListWrapper>
+          {RemoteData.hasData(markets) &&
+            markets.data.length > 0 &&
+            markets.data.slice(0, count).map(item => {
+              return <ListItem key={item.id} market={item}></ListItem>
+            })}
+          {RemoteData.is.success(markets) && markets.data.length === 0 && (
+            <NoMarketsAvailable title={'No markets available'} />
+          )}
+        </ListWrapper>
         {moreMarkets && showMoreButton}
       </ListCard>
       {RemoteData.is.loading(markets) ? <Loading message="Loading markets..." /> : null}
